+++ src/graph.jl
@@ -1,6 +1,6 @@
 mutable struct _GraphNode
     const mask::_Mask
-    neighbors::_VecMap{_GraphNode}
+    const neighbors::_VecMap{_GraphNode}
     archetype::UInt32
 end
 
@@ -33,7 +33,7 @@ function _find_node(g::_Graph, start::_GraphNode, add::Tuple{Vararg{UInt8}}, rem
     _set_mask!(g.mask, start.mask)
     for b in remove
         if !_get_bit(g.mask, b)
-            error("entity does not have component to remove")
+            throw(ArgumentError("entity does not have component to remove"))
         end
         _clear_bit!(g.mask, b)
 
@@ -48,10 +48,10 @@ function _find_node(g::_Graph, start::_GraphNode, add::Tuple{Vararg{UInt8}}, rem
     end
     for b in add
         if _get_bit(g.mask, b)
-            error("entity already has component to add, or it was added twice")
+            throw(ArgumentError("entity already has component to add, or it was added twice"))
         end
         if _get_bit(start.mask, b)
-            error("component added and removed in the same exchange operation")
+            throw(ArgumentError("component added and removed in the same exchange operation"))
         end
         _set_bit!(g.mask, b)