+++ src/query.jl
@@ -60,6 +60,7 @@ end
 
 ```
 """
+<<<<<<< HEAD
 macro Query(world_expr, comp_types_expr)
     :(Query($(esc(world_expr)), Val.($(esc(comp_types_expr)))))
 end
@@ -142,7 +143,7 @@ end
     non_exclude_types = union(comp_types, with_types)
 
     if EX === Val{true} && !isempty(without_types)
-        error("cannot use 'exclusive' together with 'without'")
+        throw(ArgumentError("cannot use 'exclusive' together with 'without'"))
     end
 
     function get_id(C)
@@ -159,7 +160,7 @@ end
     exclude_mask = EX === Val{true} ? _MaskNot(non_exclude_ids...) : _Mask(without_ids...)
     has_excluded = (length(without_ids) > 0) || (EX === Val{true})
 
-    storage_types = [_ComponentStorage{T} for T in comp_types]
+    storage_types = [_ComponentStorage{T,Vector{T}} for T in comp_types]
     storage_tuple_type = Expr(:curly, :Tuple, storage_types...)
 
     storage_exprs = Expr[:(world._storages[$(Int(i))]) for i in all_ids]